♻️refactor: RequestHeader, ResponseHeader를 하나의 공용 클래스로 수정
